moved ADDAwrapper to misc, fixed bug in crosssec.c
