--- examples/refractive_index/Ag_JHW.csv
@@ -1,170 +1,0 @@
-0.10,5.08,86.53
-0.15,3.23,59.73
-0.20,1.85,45.26
-0.25,1.23,36.37
-0.30,0.86,30.37
-0.35,0.64,26.05
-0.40,0.50,22.80
-0.45,0.40,20.26
-0.50,0.33,18.23
-0.55,0.27,16.56
-0.60,0.23,15.17
-0.65,0.20,13.98
-0.70,0.17,12.97
-0.75,0.15,12.08
-0.80,0.14,11.31
-0.85,0.13,10.63
-0.90,0.11,10.02
-0.95,0.11,9.47
-1.00,0.10,8.98
-1.05,0.09,8.53
-1.10,0.08,8.13
-1.15,0.08,7.76
-1.20,0.07,7.41
-1.25,0.07,7.10
-1.30,0.07,6.80
-1.35,0.06,6.53
-1.40,0.06,6.28
-1.45,0.06,6.04
-1.50,0.06,5.81
-1.55,0.05,5.61
-1.60,0.05,5.41
-1.65,0.05,5.22
-1.70,0.05,5.05
-1.75,0.05,4.88
-1.80,0.05,4.72
-1.85,0.05,4.57
-1.90,0.05,4.43
-1.95,0.05,4.30
-2.00,0.05,4.16
-2.05,0.05,4.04
-2.10,0.05,3.92
-2.15,0.05,3.80
-2.20,0.05,3.69
-2.25,0.05,3.59
-2.30,0.05,3.48
-2.35,0.05,3.38
-2.40,0.05,3.29
-2.45,0.05,3.20
-2.50,0.05,3.10
-2.55,0.05,3.02
-2.60,0.05,2.93
-2.65,0.05,2.85
-2.70,0.05,2.76
-2.75,0.05,2.68
-2.80,0.05,2.60
-2.85,0.05,2.53
-2.90,0.05,2.45
-2.95,0.05,2.37
-3.00,0.05,2.29
-3.05,0.05,2.22
-3.10,0.05,2.14
-3.15,0.05,2.06
-3.20,0.05,1.98
-3.25,0.05,1.90
-3.30,0.06,1.81
-3.35,0.07,1.73
-3.40,0.08,1.64
-3.45,0.09,1.56
-3.50,0.11,1.47
-3.55,0.12,1.37
-3.60,0.14,1.27
-3.65,0.16,1.16
-3.70,0.19,1.04
-3.75,0.22,0.89
-3.80,0.31,0.65
-3.85,0.62,0.44
-3.90,0.86,0.47
-3.95,1.08,0.54
-4.00,1.27,0.66
-4.05,1.35,0.78
-4.10,1.41,0.89
-4.15,1.43,1.00
-4.20,1.43,1.08
-4.25,1.42,1.14
-4.30,1.40,1.18
-4.35,1.39,1.22
-4.40,1.38,1.25
-4.45,1.37,1.27
-4.50,1.35,1.29
-4.55,1.33,1.30
-4.60,1.32,1.31
-4.65,1.31,1.32
-4.70,1.29,1.33
-4.75,1.28,1.33
-4.80,1.27,1.33
-4.85,1.26,1.33
-4.90,1.25,1.33
-4.95,1.24,1.33
-5.00,1.23,1.33
-5.05,1.22,1.32
-5.10,1.21,1.32
-5.15,1.21,1.31
-5.20,1.20,1.31
-5.25,1.19,1.30
-5.30,1.19,1.30
-5.35,1.19,1.29
-5.40,1.18,1.29
-5.45,1.18,1.29
-5.50,1.17,1.28
-5.55,1.16,1.28
-5.60,1.16,1.28
-5.65,1.15,1.28
-5.70,1.14,1.27
-5.75,1.13,1.27
-5.80,1.12,1.27
-5.85,1.11,1.26
-5.90,1.10,1.26
-5.95,1.09,1.25
-6.00,1.08,1.25
-6.05,1.07,1.24
-6.10,1.05,1.23
-6.42,0.99,1.15
-6.92,0.92,1.01
-7.42,0.91,0.87
-7.92,0.92,0.72
-8.42,1.01,0.60
-8.92,1.11,0.55
-9.42,1.18,0.56
-9.92,1.21,0.57
-10.42,1.23,0.57
-10.92,1.25,0.57
-11.42,1.26,0.58
-11.92,1.27,0.59
-12.42,1.27,0.59
-12.92,1.28,0.61
-13.42,1.28,0.63
-13.92,1.26,0.66
-14.42,1.24,0.68
-14.92,1.20,0.69
-15.42,1.16,0.69
-15.92,1.12,0.67
-16.42,1.09,0.65
-16.92,1.07,0.63
-17.42,1.06,0.61
-17.92,1.05,0.59
-18.42,1.05,0.57
-18.92,1.05,0.56
-19.42,1.05,0.55
-19.92,1.05,0.54
-20.42,1.05,0.54
-20.92,1.06,0.55
-21.42,1.05,0.57
-21.92,1.03,0.60
-22.42,0.98,0.61
-22.92,0.93,0.60
-23.42,0.89,0.57
-23.92,0.87,0.53
-24.42,0.86,0.50
-24.92,0.85,0.47
-25.42,0.85,0.45
-25.92,0.85,0.42
-26.42,0.85,0.40
-26.92,0.85,0.39
-27.42,0.86,0.38
-27.92,0.86,0.37
-28.42,0.85,0.35
-28.92,0.86,0.34
-29.42,0.86,0.33
-29.92,0.86,0.32
-30.42,0.85,0.31
