--- examples/refractive_index/Ag_Palik.csv
@@ -1,111 +1,0 @@
-0.13,13.11,53.70
-0.20,5.36,37.00
-0.30,2.45,25.10
-0.40,1.39,18.80
-0.50,0.94,15.10
-0.60,0.67,12.60
-0.70,0.56,10.70
-0.80,0.47,9.32
-0.90,0.41,8.37
-1.00,0.33,8.49
-1.10,0.25,7.67
-1.20,0.23,6.99
-1.30,0.20,6.43
-1.40,0.16,5.95
-1.50,0.15,5.50
-1.60,0.14,5.09
-1.70,0.15,4.74
-1.80,0.14,4.44
-1.90,0.14,4.15
-2.00,0.13,3.88
-2.10,0.12,3.66
-2.20,0.12,3.45
-2.30,0.13,3.25
-2.40,0.13,3.07
-2.50,0.13,2.88
-2.60,0.13,2.72
-2.70,0.14,2.56
-2.80,0.16,2.40
-2.90,0.16,2.26
-3.00,0.17,2.11
-3.10,0.17,1.95
-3.20,0.19,1.81
-3.30,0.20,1.67
-3.40,0.19,1.61
-3.50,0.21,1.44
-3.60,0.24,1.24
-3.70,0.29,0.99
-3.80,0.53,0.66
-3.90,0.93,0.50
-3.95,1.15,0.54
-4.00,1.32,0.65
-4.10,1.50,0.88
-4.20,1.52,1.08
-4.30,1.50,1.19
-4.40,1.48,1.26
-4.50,1.44,1.31
-4.60,1.40,1.33
-4.70,1.37,1.35
-4.80,1.34,1.35
-4.90,1.32,1.35
-5.00,1.30,1.35
-5.20,1.27,1.33
-5.40,1.24,1.31
-5.60,1.21,1.30
-5.80,1.17,1.29
-6.00,1.13,1.27
-6.20,1.07,1.24
-6.30,1.05,1.21
-6.40,1.03,1.18
-6.60,1.00,1.13
-6.80,0.97,1.07
-7.00,0.95,1.01
-7.20,0.94,0.95
-7.40,0.94,0.89
-7.60,0.94,0.83
-7.80,0.94,0.77
-8.00,0.96,0.71
-8.20,0.99,0.65
-8.40,1.03,0.61
-8.60,1.07,0.58
-8.80,1.11,0.56
-9.00,1.15,0.55
-9.20,1.18,0.55
-9.70,1.23,0.57
-10.00,1.24,0.57
-10.30,1.25,0.56
-10.60,1.27,0.56
-10.90,1.28,0.56
-11.20,1.29,0.57
-11.50,1.30,0.57
-11.80,1.30,0.58
-12.10,1.31,0.58
-12.50,1.32,0.59
-13.00,1.32,0.60
-13.50,1.32,0.63
-14.00,1.31,0.66
-14.50,1.28,0.68
-15.00,1.24,0.69
-15.50,1.20,0.69
-16.00,1.17,0.68
-17.00,1.12,0.64
-18.00,1.10,0.59
-19.00,1.10,0.57
-20.00,1.10,0.55
-20.50,1.11,0.55
-21.00,1.11,0.56
-21.50,1.10,0.58
-22.00,1.08,0.62
-22.50,1.03,0.62
-23.00,0.98,0.61
-23.50,0.94,0.58
-24.50,0.91,0.52
-25.50,0.89,0.47
-26.00,0.89,0.44
-26.50,0.89,0.42
-27.00,0.88,0.40
-27.50,0.85,0.62
-28.00,0.90,0.59
-28.50,0.91,0.57
-29.00,0.92,0.57
-30.00,0.93,0.54
