--- examples/refractive_index/Ag_Palik_Garcia01.csv
@@ -1,41 +1,0 @@
-1,0.328999978,8.489999704
-1.1,0.251000052,7.670000659
-1.2,0.22600001,6.990000143
-1.3,0.197999972,6.429999555
-1.4,0.163000012,5.949999701
-1.5,0.144999991,5.499999974
-1.6,0.143000004,5.090000516
-1.7,0.147999997,4.739999951
-1.8,0.140000008,4.440000104
-1.9,0.140000007,4.150000003
-2,0.130999989,3.880000094
-2.1,0.120999999,3.65999999
-2.2,0.119999999,3.450000399
-2.3,0.128999993,3.249999909
-2.4,0.129999991,3.070000184
-2.5,0.129999988,2.880000347
-2.6,0.132000011,2.719999811
-2.7,0.143999989,2.559999986
-2.8,0.157000003,2.400000144
-2.9,0.159999995,2.259999942
-3,0.172999989,2.109999874
-3.1,0.172999996,1.949999956
-3.2,0.191999972,1.81000016
-3.3,0.199999985,1.669999926
-3.4,0.18600003,1.610000038
-3.5,0.209000046,1.439999805
-3.6,0.237999978,1.240000137
-3.7,0.293999806,0.986000652
-3.8,0.514777354,0.675192153
-3.9,0.931998905,0.504000168
-4,1.32300004,0.647000105
-4.1,1.496000164,0.882000612
-4.2,1.519000026,1.08000012
-4.3,1.501999999,1.189999965
-4.4,1.476000022,1.259999792
-4.5,1.44099999,1.309999898
-4.6,1.403999971,1.330000052
-4.7,1.371999877,1.350000034
-4.8,1.34300013,1.350000041
-4.9,1.320000034,1.350000011
-5,1.298000021,1.350000032
