--- examples/refractive_index/Au_JHW.csv
@@ -1,168 +1,0 @@
-0.10,8.96,84.09
-0.15,4.02,56.47
-0.20,2.26,42.44
-0.25,1.45,33.95
-0.30,1.00,28.27
-0.35,0.74,24.20
-0.40,0.56,21.09
-0.45,0.53,18.72
-0.50,0.43,16.86
-0.55,0.35,15.30
-0.60,0.30,14.00
-0.65,0.25,12.89
-0.70,0.22,11.93
-0.75,0.19,11.08
-0.80,0.19,10.35
-0.85,0.17,9.72
-0.90,0.15,9.14
-0.95,0.13,8.61
-1.00,0.14,8.14
-1.05,0.12,7.72
-1.10,0.11,7.33
-1.15,0.10,6.97
-1.20,0.10,6.63
-1.25,0.10,6.33
-1.30,0.09,6.04
-1.35,0.09,5.77
-1.40,0.09,5.52
-1.45,0.08,5.28
-1.50,0.08,5.05
-1.55,0.08,4.83
-1.60,0.08,4.63
-1.65,0.08,4.43
-1.70,0.08,4.24
-1.75,0.08,4.05
-1.80,0.08,3.87
-1.85,0.10,3.70
-1.90,0.11,3.53
-1.95,0.12,3.37
-2.00,0.14,3.20
-2.05,0.15,3.04
-2.10,0.18,2.88
-2.15,0.21,2.72
-2.20,0.24,2.57
-2.25,0.28,2.41
-2.30,0.33,2.25
-2.35,0.40,2.05
-2.40,0.50,1.85
-2.45,0.66,1.72
-2.50,0.86,1.63
-2.55,1.05,1.53
-2.60,1.27,1.56
-2.65,1.40,1.65
-2.70,1.46,1.74
-2.75,1.46,1.78
-2.80,1.49,1.80
-2.85,1.51,1.80
-2.90,1.53,1.81
-2.95,1.55,1.82
-3.00,1.56,1.83
-3.05,1.57,1.83
-3.10,1.57,1.84
-3.15,1.57,1.83
-3.20,1.57,1.82
-3.25,1.57,1.81
-3.30,1.58,1.80
-3.35,1.58,1.79
-3.40,1.59,1.78
-3.45,1.60,1.77
-3.50,1.61,1.75
-3.55,1.63,1.75
-3.60,1.64,1.75
-3.65,1.66,1.76
-3.70,1.67,1.78
-3.75,1.67,1.80
-3.80,1.66,1.82
-3.85,1.63,1.83
-3.90,1.61,1.83
-3.95,1.59,1.83
-4.00,1.57,1.83
-4.05,1.55,1.82
-4.10,1.54,1.81
-4.15,1.52,1.81
-4.20,1.50,1.80
-4.25,1.49,1.79
-4.30,1.47,1.79
-4.35,1.45,1.79
-4.40,1.43,1.78
-4.45,1.40,1.77
-4.50,1.37,1.76
-4.55,1.35,1.74
-4.60,1.32,1.71
-4.65,1.30,1.68
-4.70,1.29,1.66
-4.75,1.28,1.63
-4.80,1.27,1.61
-4.85,1.25,1.58
-4.90,1.25,1.56
-4.95,1.24,1.53
-5.00,1.24,1.50
-5.05,1.23,1.48
-5.10,1.23,1.46
-5.15,1.23,1.44
-5.20,1.23,1.42
-5.25,1.23,1.40
-5.30,1.23,1.38
-5.35,1.23,1.36
-5.40,1.23,1.35
-5.45,1.23,1.33
-5.50,1.23,1.31
-5.55,1.23,1.30
-5.60,1.23,1.28
-5.65,1.22,1.26
-5.70,1.23,1.24
-5.75,1.23,1.23
-5.80,1.23,1.21
-5.85,1.23,1.20
-5.90,1.24,1.18
-5.95,1.24,1.17
-6.00,1.24,1.15
-6.10,1.25,1.13
-6.51,1.28,1.04
-7.01,1.36,0.96
-7.51,1.41,0.98
-8.01,1.33,0.97
-8.51,1.33,0.89
-9.01,1.32,0.84
-9.51,1.37,0.79
-10.01,1.39,0.81
-10.51,1.37,0.80
-11.01,1.36,0.76
-11.51,1.38,0.73
-12.01,1.42,0.72
-12.51,1.46,0.76
-13.01,1.46,0.83
-13.51,1.40,0.86
-14.01,1.35,0.87
-14.51,1.30,0.86
-15.01,1.27,0.84
-15.51,1.23,0.82
-16.01,1.22,0.80
-16.51,1.21,0.78
-17.01,1.20,0.76
-17.51,1.20,0.75
-18.01,1.20,0.74
-18.51,1.21,0.74
-19.01,1.22,0.75
-19.51,1.22,0.80
-20.01,1.19,0.83
-20.51,1.14,0.86
-21.01,1.08,0.88
-21.51,1.01,0.88
-22.01,0.94,0.86
-22.51,0.88,0.82
-23.01,0.84,0.76
-23.51,0.81,0.70
-24.01,0.80,0.65
-24.51,0.80,0.60
-25.01,0.81,0.56
-25.51,0.83,0.53
-26.01,0.84,0.51
-26.51,0.85,0.49
-27.01,0.86,0.48
-27.51,0.86,0.47
-28.01,0.87,0.47
-28.51,0.87,0.46
-29.01,0.87,0.47
-29.51,0.86,0.46
-30.01,0.84,0.45
