--- examples/refractive_index/Au_Palik.csv
@@ -1,95 +1,0 @@
-0.13,12.24,54.70
-0.15,9.02,47.60
-0.20,5.42,37.50
-0.25,3.75,30.50
-0.30,2.75,25.40
-0.40,1.73,19.20
-0.50,1.21,15.50
-0.60,0.90,13.00
-0.70,0.70,11.20
-0.80,0.56,9.81
-0.90,0.45,8.77
-1.00,0.37,8.77
-1.10,0.31,7.93
-1.20,0.27,7.07
-1.30,0.24,6.47
-1.40,0.21,5.88
-1.50,0.19,5.39
-1.60,0.17,4.86
-1.70,0.16,4.35
-1.80,0.16,3.80
-1.90,0.17,3.15
-2.20,0.31,2.88
-2.30,0.40,2.54
-2.40,0.61,2.12
-2.50,0.92,1.84
-2.60,1.24,1.80
-2.70,1.43,1.85
-2.80,1.56,1.90
-2.90,1.62,1.94
-3.00,1.64,1.96
-3.10,1.66,1.96
-3.20,1.67,1.94
-3.30,1.70,1.91
-3.40,1.72,1.86
-3.50,1.74,1.85
-3.60,1.77,1.85
-3.70,1.80,1.86
-3.80,1.82,1.88
-3.90,1.84,1.90
-4.00,1.83,1.92
-4.10,1.81,1.92
-4.20,1.78,1.92
-4.30,1.74,1.90
-4.40,1.69,1.88
-4.50,1.65,1.85
-4.60,1.60,1.82
-4.70,1.55,1.78
-4.80,1.50,1.75
-4.90,1.49,1.70
-5.00,1.48,1.64
-5.10,1.48,1.59
-5.20,1.47,1.55
-5.30,1.46,1.51
-5.40,1.45,1.48
-5.50,1.45,1.44
-5.60,1.44,1.42
-5.70,1.44,1.39
-5.80,1.43,1.36
-5.90,1.43,1.33
-6.00,1.42,1.31
-6.53,1.43,1.14
-7.09,1.50,1.07
-8.00,1.45,1.11
-8.98,1.35,1.06
-9.47,1.31,1.03
-10.00,1.28,0.99
-10.51,1.25,0.95
-11.07,1.23,0.91
-11.59,1.22,0.88
-12.04,1.21,0.85
-12.52,1.19,0.83
-13.62,1.10,0.80
-14.42,1.05,0.78
-15.50,1.03,0.75
-16.31,1.03,0.71
-17.22,1.05,0.69
-18.23,1.11,0.69
-19.68,1.15,0.75
-20.00,1.15,0.77
-21.01,1.09,0.83
-21.75,1.02,0.85
-22.54,0.95,0.84
-22.96,0.92,0.83
-23.84,0.87,0.74
-24.31,0.86,0.70
-24.80,0.85,0.65
-25.30,0.85,0.60
-25.83,0.85,0.57
-26.00,0.90,0.68
-26.38,0.86,0.55
-27.00,0.90,0.64
-28.00,0.90,0.61
-29.00,0.91,0.60
-29.50,0.90,0.60
-30.00,0.89,0.60
