--- examples/refractive_index/Cu_JHW.csv
@@ -1,169 +1,0 @@
-0.10,11.18,85.47
-0.15,5.13,58.62
-0.20,2.97,44.35
-0.25,1.92,35.62
-0.30,1.33,29.73
-0.35,0.98,25.48
-0.40,0.75,22.28
-0.45,0.59,19.78
-0.50,0.48,17.78
-0.55,0.40,16.14
-0.60,0.33,14.76
-0.65,0.28,13.59
-0.70,0.24,12.59
-0.75,0.21,11.72
-0.80,0.19,10.95
-0.85,0.16,10.27
-0.90,0.15,9.66
-0.95,0.13,9.11
-1.00,0.12,8.62
-1.05,0.11,8.17
-1.10,0.10,7.76
-1.15,0.09,7.38
-1.20,0.08,7.03
-1.25,0.08,6.70
-1.30,0.07,6.40
-1.35,0.07,6.12
-1.40,0.07,5.85
-1.45,0.06,5.60
-1.50,0.06,5.36
-1.55,0.06,5.13
-1.60,0.06,4.92
-1.65,0.06,4.71
-1.70,0.06,4.51
-1.75,0.07,4.31
-1.80,0.07,4.12
-1.85,0.08,3.94
-1.90,0.09,3.76
-1.95,0.09,3.58
-2.00,0.09,3.38
-2.05,0.10,3.18
-2.10,0.13,2.92
-2.15,0.25,2.62
-2.20,0.52,2.41
-2.25,0.79,2.36
-2.30,0.97,2.41
-2.35,1.05,2.47
-2.40,1.09,2.50
-2.45,1.10,2.50
-2.50,1.12,2.48
-2.55,1.13,2.47
-2.60,1.13,2.45
-2.65,1.14,2.41
-2.70,1.15,2.37
-2.75,1.15,2.34
-2.80,1.17,2.30
-2.85,1.18,2.27
-2.90,1.19,2.24
-2.95,1.19,2.20
-3.00,1.20,2.16
-3.05,1.21,2.11
-3.10,1.24,2.08
-3.15,1.26,2.05
-3.20,1.28,2.04
-3.25,1.28,2.01
-3.30,1.29,1.99
-3.35,1.31,1.96
-3.40,1.32,1.94
-3.45,1.34,1.92
-3.50,1.35,1.90
-3.55,1.36,1.89
-3.60,1.37,1.87
-3.65,1.37,1.86
-3.70,1.38,1.85
-3.75,1.38,1.84
-3.80,1.38,1.81
-3.85,1.38,1.79
-3.90,1.39,1.77
-3.95,1.40,1.75
-4.00,1.40,1.73
-4.05,1.41,1.72
-4.10,1.42,1.70
-4.15,1.44,1.68
-4.20,1.46,1.66
-4.25,1.48,1.65
-4.30,1.51,1.65
-4.35,1.53,1.66
-4.40,1.54,1.66
-4.45,1.55,1.67
-4.50,1.56,1.68
-4.55,1.57,1.69
-4.60,1.58,1.70
-4.65,1.58,1.72
-4.70,1.58,1.73
-4.75,1.58,1.75
-4.80,1.57,1.77
-4.85,1.56,1.79
-4.90,1.55,1.81
-4.95,1.52,1.83
-5.00,1.49,1.84
-5.05,1.46,1.84
-5.10,1.44,1.84
-5.15,1.42,1.85
-5.20,1.39,1.85
-5.25,1.36,1.85
-5.30,1.33,1.84
-5.35,1.30,1.83
-5.40,1.28,1.82
-5.45,1.25,1.81
-5.50,1.23,1.80
-5.55,1.20,1.79
-5.60,1.18,1.77
-5.65,1.16,1.76
-5.70,1.13,1.74
-5.75,1.11,1.72
-5.80,1.09,1.70
-5.85,1.07,1.68
-5.90,1.05,1.65
-5.95,1.04,1.63
-6.00,1.02,1.60
-6.05,1.01,1.57
-6.10,1.01,1.55
-6.42,0.98,1.41
-6.92,0.98,1.23
-7.42,1.02,1.12
-7.92,1.04,1.05
-8.42,1.05,1.00
-8.92,1.04,0.94
-9.42,1.05,0.88
-9.92,1.05,0.83
-10.42,1.07,0.79
-10.92,1.09,0.76
-11.42,1.10,0.74
-11.92,1.11,0.74
-12.42,1.11,0.73
-12.92,1.11,0.73
-13.42,1.10,0.73
-13.92,1.08,0.73
-14.42,1.06,0.72
-14.92,1.04,0.72
-15.42,1.01,0.71
-15.92,0.98,0.69
-16.42,0.96,0.66
-16.92,0.94,0.64
-17.42,0.92,0.61
-17.92,0.91,0.58
-18.42,0.90,0.55
-18.92,0.90,0.52
-19.42,0.90,0.49
-19.92,0.90,0.47
-20.42,0.90,0.44
-20.92,0.91,0.42
-21.42,0.92,0.40
-21.92,0.93,0.39
-22.42,0.94,0.38
-22.92,0.96,0.37
-23.42,0.97,0.37
-23.92,0.98,0.38
-24.42,0.98,0.39
-24.92,0.98,0.40
-25.42,0.97,0.41
-25.92,0.95,0.41
-26.42,0.92,0.41
-26.92,0.90,0.39
-27.42,0.89,0.38
-27.92,0.88,0.36
-28.42,0.87,0.34
-28.92,0.87,0.32
-29.42,0.87,0.30
-29.92,0.87,0.28
