--- examples/refractive_index/Cu_Palik.csv
@@ -1,66 +1,0 @@
-0.13,10.80,47.50
-0.30,2.59,22.20
-0.50,1.15,13.20
-0.72,0.70,9.16
-0.88,0.55,7.53
-0.98,0.50,6.78
-1.00,0.43,8.46
-1.50,0.26,5.26
-1.70,0.22,4.43
-1.75,0.21,4.24
-1.80,0.21,4.05
-1.85,0.22,3.86
-1.90,0.21,3.67
-2.00,0.27,3.24
-2.10,0.47,2.81
-2.20,0.83,2.60
-2.30,1.04,2.59
-2.40,1.12,2.60
-2.60,1.15,2.50
-2.80,1.17,2.36
-3.00,1.18,2.21
-3.20,1.18,2.21
-3.40,1.27,1.95
-3.60,1.31,1.87
-3.80,1.34,1.81
-4.00,1.34,1.72
-4.20,1.42,1.64
-4.40,1.49,1.64
-4.60,1.52,1.67
-4.80,1.53,1.71
-5.00,1.47,1.78
-5.20,1.38,1.80
-5.40,1.28,1.78
-5.60,1.18,1.74
-5.80,1.10,1.67
-6.00,1.04,1.59
-6.50,0.96,1.37
-7.00,0.97,1.20
-7.50,1.01,1.09
-8.00,1.03,1.03
-8.50,1.03,0.98
-9.00,1.03,0.92
-9.50,1.03,0.87
-10.00,1.04,0.82
-11.00,1.07,0.75
-12.00,1.09,0.73
-13.00,1.06,0.72
-14.00,1.08,0.72
-14.50,1.03,0.72
-15.00,1.01,0.71
-15.50,0.98,0.70
-16.00,0.95,0.67
-17.00,0.91,0.62
-18.00,0.89,0.56
-19.00,0.88,0.51
-20.00,0.88,0.46
-21.00,0.90,0.41
-22.00,0.92,0.38
-23.00,0.94,0.37
-24.00,0.96,0.37
-25.00,0.96,0.40
-26.00,0.92,0.40
-27.00,0.88,0.38
-28.00,0.86,0.35
-29.00,0.85,0.30
-30.00,0.86,0.26
