+++ misc/ADDAwrapper/refractive_index/Ag_JC.csv
@@ -1,0 +1,49 @@
+0.64,0.24,14.08
+0.77,0.15,11.85
+0.89,0.13,10.10
+1.02,0.09,8.83
+1.14,0.04,7.80
+1.26,0.04,6.99
+1.39,0.04,6.31
+1.51,0.04,5.73
+1.64,0.03,5.24
+1.76,0.04,4.84
+1.88,0.05,4.48
+2.01,0.06,4.15
+2.13,0.05,3.86
+2.26,0.06,3.59
+2.38,0.05,3.32
+2.5,0.05,3.09
+2.63,0.05,2.87
+2.75,0.04,2.66
+2.88,0.04,2.46
+3,0.05,2.28
+3.12,0.05,2.07
+3.25,0.05,1.86
+3.37,0.07,1.66
+3.5,0.10,1.42
+3.62,0.14,1.14
+3.74,0.17,0.83
+3.87,0.81,0.39
+3.99,1.13,0.62
+4.12,1.34,0.96
+4.24,1.39,1.16
+4.36,1.41,1.26
+4.49,1.41,1.33
+4.61,1.38,1.37
+4.74,1.35,1.39
+4.86,1.33,1.39
+4.98,1.31,1.39
+5.11,1.30,1.38
+5.23,1.28,1.37
+5.36,1.28,1.36
+5.48,1.26,1.34
+5.6,1.25,1.34
+5.73,1.22,1.34
+5.85,1.20,1.33
+5.98,1.18,1.31
+6.1,1.15,1.30
+6.22,1.14,1.28
+6.35,1.12,1.26
+6.47,1.10,1.23
+6.6,1.07,1.21
