+++ misc/ADDAwrapper/refractive_index/Ag_Palik.csv
@@ -1,0 +1,110 @@
+0.13,13.11,53.7
+0.2,5.36,37
+0.3,2.45,25.1
+0.4,1.39,18.8
+0.5,0.94,15.1
+0.6,0.67,12.6
+0.7,0.56,10.7
+0.8,0.47,9.32
+0.9,0.41,8.37
+1,0.33,8.49
+1.1,0.25,7.67
+1.2,0.23,6.99
+1.3,0.2,6.43
+1.4,0.16,5.95
+1.5,0.15,5.5
+1.6,0.14,5.09
+1.7,0.15,4.74
+1.8,0.14,4.44
+1.9,0.14,4.15
+2,0.13,3.88
+2.1,0.12,3.66
+2.2,0.12,3.45
+2.3,0.13,3.25
+2.4,0.13,3.07
+2.5,0.13,2.88
+2.6,0.13,2.72
+2.7,0.14,2.56
+2.8,0.16,2.4
+2.9,0.16,2.26
+3,0.17,2.11
+3.1,0.17,1.95
+3.2,0.19,1.81
+3.3,0.2,1.67
+3.4,0.19,1.61
+3.5,0.21,1.44
+3.6,0.24,1.24
+3.7,0.29,0.99
+3.8,0.53,0.66
+3.9,0.93,0.5
+4,1.32,0.65
+4.1,1.5,0.88
+4.2,1.52,1.08
+4.3,1.5,1.19
+4.4,1.48,1.26
+4.5,1.44,1.31
+4.6,1.4,1.33
+4.7,1.37,1.35
+4.8,1.34,1.35
+4.9,1.32,1.35
+5,1.3,1.35
+5.2,1.27,1.33
+5.4,1.24,1.31
+5.6,1.21,1.3
+5.8,1.17,1.29
+6,1.13,1.27
+6.2,1.07,1.24
+6.3,1.05,1.21
+6.4,1.03,1.18
+6.6,1,1.13
+6.8,0.97,1.07
+7,0.95,1.01
+7.2,0.94,0.95
+7.4,0.94,0.89
+7.6,0.94,0.83
+7.8,0.94,0.77
+8,0.96,0.71
+8.2,0.99,0.65
+8.4,1.03,0.61
+8.6,1.07,0.58
+8.8,1.11,0.56
+9,1.15,0.55
+9.2,1.18,0.55
+9.7,1.23,0.57
+10,1.24,0.57
+10.3,1.25,0.56
+10.6,1.27,0.56
+10.9,1.28,0.56
+11.2,1.29,0.57
+11.5,1.3,0.57
+11.8,1.3,0.58
+12.1,1.31,0.58
+12.5,1.32,0.59
+13,1.32,0.6
+13.5,1.32,0.63
+14,1.31,0.66
+14.5,1.28,0.68
+15,1.24,0.69
+15.5,1.2,0.69
+16,1.17,0.68
+17,1.12,0.64
+18,1.1,0.59
+19,1.1,0.57
+20,1.1,0.55
+20.5,1.11,0.55
+21,1.11,0.56
+21.5,1.1,0.58
+22,1.08,0.62
+22.5,1.03,0.62
+23,0.98,0.61
+23.5,0.94,0.58
+24.5,0.91,0.52
+25.5,0.89,0.47
+26,0.89,0.44
+26.5,0.89,0.42
+27,0.88,0.4
+27.5,0.85,0.62
+28,0.9,0.59
+28.5,0.91,0.57
+29,0.92,0.57
+30,0.93,0.54
