+++ misc/ADDAwrapper/refractive_index/Ag_Palik_Garcia.csv
@@ -79,12 +79,3 @@
 4.9,1.320000034,1.350000011
 4.95,1.309022425,1.349976922
 5,1.298000021,1.350000032
-5.05,1.289821919,1.345020986
-5.1,1.281596393,1.340027972
-5.15,1.273322668,1.335020987
-5.2,1.264999954,1.330000024
-5.25,1.258301774,1.325025915
-5.3,1.2515694,1.320034634
-5.35,1.244802325,1.315026022
-5.4,1.238000022,1.309999929
-5.45,1.230551448,1.307490985
