+++ misc/ADDAwrapper/refractive_index/Au_JC.csv
@@ -1,0 +1,49 @@
+0.64,0.92,13.78
+0.77,0.56,11.21
+0.89,0.43,9.52
+1.02,0.35,8.15
+1.14,0.27,7.15
+1.26,0.22,6.35
+1.39,0.17,5.66
+1.51,0.16,5.08
+1.64,0.14,4.54
+1.76,0.13,4.10
+1.88,0.14,3.70
+2.01,0.21,3.27
+2.13,0.29,2.86
+2.26,0.43,2.46
+2.38,0.62,2.08
+2.5,1.04,1.83
+2.63,1.31,1.85
+2.75,1.38,1.91
+2.88,1.45,1.95
+3,1.46,1.96
+3.12,1.47,1.95
+3.25,1.46,1.93
+3.37,1.48,1.90
+3.5,1.50,1.87
+3.62,1.48,1.87
+3.74,1.48,1.88
+3.87,1.54,1.90
+3.99,1.53,1.89
+4.12,1.53,1.89
+4.24,1.49,1.88
+4.36,1.47,1.87
+4.49,1.43,1.85
+4.61,1.38,1.80
+4.74,1.35,1.75
+4.86,1.33,1.69
+4.98,1.33,1.63
+5.11,1.32,1.58
+5.23,1.32,1.54
+5.36,1.30,1.50
+5.48,1.31,1.46
+5.6,1.30,1.43
+5.73,1.30,1.39
+5.85,1.30,1.35
+5.98,1.30,1.30
+6.1,1.33,1.28
+6.22,1.33,1.25
+6.35,1.34,1.23
+6.47,1.32,1.20
+6.6,1.28,1.19
