+++ misc/ADDAwrapper/refractive_index/Cu_JC.csv
@@ -1,0 +1,49 @@
+0.64,1.09,13.43
+0.77,0.76,11.12
+0.89,0.60,9.44
+1.02,0.48,8.25
+1.14,0.36,7.22
+1.26,0.32,6.42
+1.39,0.30,5.77
+1.51,0.26,5.18
+1.64,0.24,4.67
+1.76,0.21,4.21
+1.88,0.22,3.75
+2.01,0.30,3.21
+2.13,0.70,2.70
+2.26,1.02,2.58
+2.38,1.18,2.61
+2.50,1.22,2.56
+2.63,1.25,2.48
+2.75,1.24,2.40
+2.88,1.25,2.31
+3.00,1.28,2.21
+3.12,1.32,2.12
+3.25,1.33,2.05
+3.37,1.36,1.98
+3.50,1.37,1.92
+3.62,1.36,1.86
+3.74,1.34,1.82
+3.87,1.38,1.78
+3.99,1.38,1.73
+4.12,1.40,1.68
+4.24,1.42,1.63
+4.36,1.45,1.63
+4.49,1.46,1.65
+4.61,1.45,1.67
+4.74,1.41,1.69
+4.86,1.41,1.74
+4.98,1.37,1.78
+5.11,1.34,1.80
+5.23,1.28,1.80
+5.36,1.23,1.79
+5.48,1.18,1.77
+5.60,1.13,1.74
+5.73,1.08,1.70
+5.85,1.04,1.65
+5.98,1.01,1.60
+6.10,0.99,1.55
+6.22,0.98,1.49
+6.35,0.97,1.44
+6.47,0.95,1.39
+6.60,0.94,1.34
